+++ src/CustomerInformation.php
@@ -16,7 +16,7 @@ use DateTime;
  * Customer information.
  *
  * @author  Remco Tolsma
- * @version 2.1.10
+ * @version 2.2.4
  * @since   2.0.2
  */
 class CustomerInformation {
@@ -87,7 +87,7 @@ class CustomerInformation {
 			throw new \InvalidArgumentException(
 				\sprintf(
 					'Gender "%s" must be equal to `null`, "F" or "M".',
-					$gender
+					\strval( $gender )
 				)
 			);
 		}