+++ src/Gateway.php
@@ -14,13 +14,12 @@ use Pronamic\WordPress\Money\TaxedMoney;
 use Pronamic\WordPress\Pay\Core\Gateway as Core_Gateway;
 use Pronamic\WordPress\Pay\Core\PaymentMethods;
 use Pronamic\WordPress\Pay\Payments\Payment;
-use Pronamic\WordPress\Pay\Payments\PaymentStatus;
 
 /**
  * Gateway
  *
  * @author  Remco Tolsma
- * @version 2.2.3
+ * @version 2.2.4
  * @since   1.0.0
  */
 class Gateway extends Core_Gateway {
@@ -221,11 +220,7 @@ class Gateway extends Core_Gateway {
 		// Announce order.
 		$response = $this->client->order_announce( $this->config, $order );
 
-		// Validate.
-		if ( ! $response->is_valid( $this->config->signing_key ) ) {
-			throw new \Exception( 'Could not validate OmniKassa 2.0 response signature with signing key.' );
-		}
-
+		$payment->set_transaction_id( $response->get_omnikassa_order_id() );
 		$payment->set_action_url( $response->get_redirect_url() );
 	}
 
@@ -248,7 +243,7 @@ class Gateway extends Core_Gateway {
 		$note_values = array(
 			'order_id'  => $parameters->get_order_id(),
 			'status'    => $parameters->get_status(),
-			'signature' => $parameters->get_signature(),
+			'signature' => \strval( $parameters->get_signature() ),
 			'valid'     => $parameters->is_valid( $this->config->signing_key ) ? 'true' : 'false',
 		);
 
@@ -323,31 +318,12 @@ class Gateway extends Core_Gateway {
 
 				$payment = \get_pronamic_payment_by_transaction_id( $order_result->get_omnikassa_order_id() );
 
-				if ( empty( $payment ) ) {
-					// Get last payment, unless payment has expired, then get payments in chronological order.
-					$order = ( PaymentStatus::EXPIRED === $pronamic_status ? 'ASC' : 'DESC' );
-
-					$args = array(
-						'order'      => $order,
-						'meta_query' => array(
-							array(
-								'key'     => '_pronamic_payment_transaction_id',
-								'compare' => 'NOT EXISTS',
-							),
-						),
-					);
-
-					$payment = \get_pronamic_payment_by_meta( '_pronamic_payment_omnikassa_2_merchant_order_id', $order_result->get_merchant_order_id(), $args );
-				}
-
 				// Log webhook request.
 				\do_action( 'pronamic_pay_webhook_log_payment', $payment );
 
 				if ( empty( $payment ) ) {
 					continue;
 				}
-
-				$payment->set_transaction_id( $order_result->get_omnikassa_order_id() );
 
 				if ( null !== $pronamic_status ) {
 					$payment->set_status( $pronamic_status );